Move state conversion to separate file
